requirements.txt
Django==2.2.14
django-ipware==3.0.0
djangorestframework==3.11.0
Markdown==3.2.2
